+++ lib/openapi_first/definition.rb
@@ -45,7 +45,7 @@ module OpenapiFirst
     # @param [Boolean] raise_error Whether to raise an error if validation fails.
     # @return [ValidatedRequest] The validated request object.
     def validate_request(request, raise_error: false)
-      route = @router.match(request.request_method, request.path, content_type: request.content_type)
+      route = @router.match(request.request_method, resolve_path(request), content_type: request.content_type)
       if route.error
         ValidatedRequest.new(request, error: route.error)
       else
@@ -62,7 +62,8 @@ module OpenapiFirst
     # @param raise_error [Boolean] Whethir to raise an error if validation fails.
     # @return [ValidatedResponse] The validated response object.
     def validate_response(rack_request, rack_response, raise_error: false)
-      route = @router.match(rack_request.request_method, rack_request.path, content_type: rack_request.content_type)
+      route = @router.match(rack_request.request_method, resolve_path(rack_request),
+                            content_type: rack_request.content_type)
       return if route.error # Skip response validation for unknown requests
 
       response_match = route.match_response(status: rack_response.status, content_type: rack_response.content_type)
@@ -75,6 +76,12 @@ module OpenapiFirst
         @config.hooks[:after_response_validation]&.each { |hook| hook.call(validated, rack_request, self) }
         raise validated.error.exception(validated) if raise_error && validated.invalid?
       end
+    end
+
+    private
+
+    def resolve_path(rack_request)
+      rack_request.env[PATH] || rack_request.path
     end
   end
 end